requirements.txt
django
python-decouple
django-extensions
django-admin-extra-buttons
more-itertools
dj-database-url
gunicorn
psycopg2

model-bakery
ipdb
faker
flake8
black
isort
pre-commit
